# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Daily Creative API
  version: 0.1.0
  description: |
    Contract-first OpenAPI definition for Daily Creative.

    Version-one feature routes use the `/api/v1` prefix.
    Process health probes remain at the root (`/health/live`, `/health/ready`)
    and are not versioned under `/api/v1`.

    Authenticated feature endpoints use the `bearerAuth` security scheme
    (Descope-issued JWT). Health probes are unauthenticated.
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/v1/users/{username}:
    get:
      operationId: getPublicUser
      tags:
        - users
      summary: Public profile by username
      description: |
        Returns a public-safe projection of an active user with a completed profile.
        Available to guests and authenticated users. When authenticated,
        `is_self` is true when the viewer is the profile owner.
        Excludes email, Descope subject, preferences, moderation internals,
        Drafts, and private session analytics. Incomplete, suspended, and
        deleted profiles return 404.
        When the authenticated viewer has a block relationship with the
        profile owner in either direction, the profile is unavailable (404).
      parameters:
        - name: username
          in: path
          required: true
          description: Public username (case-insensitive lookup).
          schema:
            type: string
            minLength: 3
            maxLength: 30
          example: sketchy_matt
        - name: creative_type
          in: query
          required: true
          description: Creative type for submission counts and streak stats.
          schema:
            $ref: "#/components/schemas/CreativeType"
          example: sketch
      responses:
        "200":
          description: Public profile
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"
              example:
                id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
                username: sketchy_matt
                display_name: Matt
                bio: Sketching every day.
                avatar_url: null
                submission_count: 12
                current_streak: 5
                is_self: false
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/users/{username}/submissions:
    get:
      operationId: getUserSubmissions
      tags:
        - users
      summary: Public profile submission history
      description: |
        Returns a reverse-chronological page of published Submissions for the
        given public username, ordered by `published_at DESC, id DESC`.
        Soft-deleted, hidden, removed, and non-public content is excluded.
        Available to guests and authenticated users. When authenticated,
        `is_owner` and `viewer_has_liked` reflect the viewer.
        Incomplete, suspended, and deleted profiles return 404.
        Pass an opaque `cursor` from a previous page to continue keyset
        pagination without duplicates or skips under concurrent inserts.
      parameters:
        - name: username
          in: path
          required: true
          description: Public username (case-insensitive lookup).
          schema:
            type: string
            minLength: 3
            maxLength: 30
          example: sketchy_matt
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous page, or omit for the first page.
          schema:
            type:
              - string
              - "null"
          example: null
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return (1–50). Defaults to 20.
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          example: 20
        - name: creative_type
          in: query
          required: true
          description: Creative type of submissions to return.
          schema:
            $ref: "#/components/schemas/CreativeType"
          example: sketch
      responses:
        "200":
          description: Profile submission page
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentFeed"
              examples:
                empty:
                  summary: Empty profile history
                  value:
                    items: []
                    next_cursor: null
                populated:
                  summary: Populated reverse-chronological page
                  value:
                    items:
                      - id: d4e5f6a7-b8c9-0123-def0-234567890123
                        image_url: https://storage.example/display-signed
                        thumbnail_url: https://storage.example/thumb-signed
                        user:
                          id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
                          username: sketchy_matt
                          display_name: Matt
                          avatar_url: null
                        prompt:
                          id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                          prompt_date: "2026-07-18"
                          word_1: Chocolate
                          word_2: Coffee
                          word_3: Banana
                        timer_mode: countdown
                        timer_seconds: 300
                        caption_preview: A quiet coffee and banana sketch.
                        like_count: 0
                        reflection_count: 0
                        viewer_has_liked: false
                        is_owner: false
                        published_at: "2026-07-18T20:12:00Z"
                    next_cursor: null
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/submissions:
    post:
      operationId: createSubmission
      tags:
        - submissions
      summary: Publish a Submission
      description: |
        Atomically creates a public Submission from a ready Upload and owned
        creative Session. Requires a completed profile. Pass an `Idempotency-Key`
        header to safely retry without duplicating Submissions.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
            example:
              creative_type: sketch
              session_id: b2c3d4e5-f6a7-8901-bcde-f12345678901
              content:
                creative_type: sketch
                upload_id: c3d4e5f6-a7b8-9012-cdef-123456789012
                caption: A quiet coffee and banana sketch.
      responses:
        "201":
          description: Submission published
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
              example:
                id: d4e5f6a7-b8c9-0123-def0-234567890123
                creative_type: sketch
                caption: A quiet coffee and banana sketch.
                visibility: public
                status: published
                timer_mode: countdown
                timer_seconds: 300
                like_count: 0
                reflection_count: 0
                viewer_has_liked: false
                is_owner: true
                image_url: https://storage.example/display-signed
                thumbnail_url: https://storage.example/thumb-signed
                user:
                  id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
                  username: sketchy_matt
                  display_name: Matt
                  avatar_url: null
                prompt:
                  id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                  prompt_date: "2026-07-18"
                  word_1: Chocolate
                  word_2: Coffee
                  word_3: Banana
                session_id: b2c3d4e5-f6a7-8901-bcde-f12345678901
                upload_id: c3d4e5f6-a7b8-9012-cdef-123456789012
                body: null
                published_at: "2026-07-18T20:12:00Z"
                created_at: "2026-07-18T20:12:00Z"
                updated_at: "2026-07-18T20:12:00Z"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/submissions/{submission_id}:
    get:
      operationId: getSubmission
      tags:
        - submissions
      summary: Get a Submission
      description: |
        Returns a published Submission. Available to guests and authenticated
        users. Soft-deleted, hidden, removed, and suspended/deleted-author
        Submissions are not returned. When authenticated, `is_owner` and
        `viewer_has_liked` reflect the viewer.
      parameters:
        - name: submission_id
          in: path
          required: true
          description: Submission UUID.
          schema:
            $ref: "#/components/schemas/UUID"
          example: d4e5f6a7-b8c9-0123-def0-234567890123
      responses:
        "200":
          description: Submission
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
              example:
                id: d4e5f6a7-b8c9-0123-def0-234567890123
                creative_type: sketch
                caption: A quiet coffee and banana sketch.
                visibility: public
                status: published
                timer_mode: countdown
                timer_seconds: 300
                like_count: 0
                reflection_count: 0
                viewer_has_liked: false
                is_owner: false
                image_url: https://storage.example/display-signed
                thumbnail_url: https://storage.example/thumb-signed
                user:
                  id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
                  username: sketchy_matt
                  display_name: Matt
                  avatar_url: null
                prompt:
                  id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                  prompt_date: "2026-07-18"
                  word_1: Chocolate
                  word_2: Coffee
                  word_3: Banana
                session_id: b2c3d4e5-f6a7-8901-bcde-f12345678901
                upload_id: c3d4e5f6-a7b8-9012-cdef-123456789012
                body: null
                published_at: "2026-07-18T20:12:00Z"
                created_at: "2026-07-18T20:12:00Z"
                updated_at: "2026-07-18T20:12:00Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteSubmission
      tags:
        - submissions
      summary: Delete own Submission
      description: |
        Soft-deletes a Submission owned by the authenticated user. The
        Submission is hidden from public GET, feed, and profile queries.
      security:
        - bearerAuth: []
      parameters:
        - name: submission_id
          in: path
          required: true
          description: Submission UUID.
          schema:
            $ref: "#/components/schemas/UUID"
          example: d4e5f6a7-b8c9-0123-def0-234567890123
      responses:
        "204":
          description: Submission deleted
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Descope-issued access/session JWT sent as Authorization Bearer token.
  headers:
    RequestId:
      description: |
        Opaque request correlation identifier. Clients may supply
        `X-Request-ID`; otherwise the server generates one. The same
        value is returned on every response and embedded in error bodies
        as `error.request_id`.
      required: true
      schema:
        type: string
        format: uuid
      example: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-generated key that makes a write operation safely retryable.
        Replaying the same key with the same request body returns the original
        response. Replaying with a different body returns `idempotency_key_conflict`.
      schema:
        type: string
        minLength: 1
        maxLength: 255
      example: sketch-session-create-7f3a9c2e
  responses:
    Unauthenticated:
      description: Missing or invalid bearer token
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthenticated
              message: Authentication is required.
              details: {}
              request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    Forbidden:
      description: Authenticated but not permitted (suspended or unavailable account)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            account_suspended:
              summary: Suspended account
              value:
                error:
                  code: account_suspended
                  message: This account has been suspended.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            account_unavailable:
              summary: Deleted or pending deletion
              value:
                error:
                  code: account_unavailable
                  message: This account is no longer available.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            profile_incomplete:
              summary: Profile incomplete
              value:
                error:
                  code: profile_incomplete
                  message: Complete your profile before publishing.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            reflection_forbidden:
              summary: Not the Reflection author
              value:
                error:
                  code: reflection_forbidden
                  message: You can only delete your own reflection.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    Conflict:
      description: Request conflicts with existing state
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            username_taken:
              summary: Username already taken
              value:
                error:
                  code: username_taken
                  message: That username is already taken.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            idempotency_key_conflict:
              summary: Idempotency key reused with a different body
              value:
                error:
                  code: idempotency_key_conflict
                  message: This idempotency key was already used with a different request.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            upload_already_consumed:
              summary: Upload already consumed
              value:
                error:
                  code: upload_already_consumed
                  message: This upload has already been used.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            session_already_submitted:
              summary: Sketch Session already has a Submission
              value:
                error:
                  code: session_already_submitted
                  message: This sketch session already has a published submission.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    NotFound:
      description: Requested resource was not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            user_not_found:
              summary: Public profile not found
              value:
                error:
                  code: user_not_found
                  message: The requested profile could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            prompt_not_found:
              summary: Daily Prompt not found
              value:
                error:
                  code: prompt_not_found
                  message: Today's prompt is not available yet.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            session_not_found:
              summary: Sketch Session not found
              value:
                error:
                  code: session_not_found
                  message: The requested sketch session could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            upload_not_found:
              summary: Upload not found
              value:
                error:
                  code: upload_not_found
                  message: The requested upload could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            submission_not_found:
              summary: Submission not found
              value:
                error:
                  code: submission_not_found
                  message: The requested sketch could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            reflection_not_found:
              summary: Reflection not found
              value:
                error:
                  code: reflection_not_found
                  message: The requested reflection could not be found.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    ValidationError:
      description: Request body failed validation or domain rules
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            username_invalid:
              summary: Invalid username format
              value:
                error:
                  code: username_invalid
                  message: Usernames must be 3–30 characters and use only letters, numbers, and underscores.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            username_reserved:
              summary: Reserved username
              value:
                error:
                  code: username_reserved
                  message: That username is reserved.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            invalid_timer_preference:
              summary: Invalid timer preference combination
              value:
                error:
                  code: invalid_timer_preference
                  message: Timer preference mode and seconds are inconsistent.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            invalid_timer_selection:
              summary: Invalid Sketch Session timer selection
              value:
                error:
                  code: invalid_timer_selection
                  message: Timer mode and selected seconds are inconsistent.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            invalid_session_transition:
              summary: Invalid Sketch Session lifecycle transition
              value:
                error:
                  code: invalid_session_transition
                  message: That lifecycle event is not valid for the current session state.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            upload_not_ready:
              summary: Upload not ready for consumption
              value:
                error:
                  code: upload_not_ready
                  message: This upload is not ready to publish yet.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            invalid_image:
              summary: Image could not be decoded or validated
              value:
                error:
                  code: invalid_image
                  message: The uploaded file is not a valid image.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            image_too_large:
              summary: Image exceeds size limit
              value:
                error:
                  code: image_too_large
                  message: That image is too large to upload.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            unsupported_media_type:
              summary: Unsupported image content type
              value:
                error:
                  code: unsupported_media_type
                  message: That image type is not supported.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            object_missing:
              summary: Storage object missing after direct upload
              value:
                error:
                  code: object_missing
                  message: The uploaded image could not be found. Please try again.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            validation_error:
              summary: Schema validation failure
              value:
                error:
                  code: validation_error
                  message: The request could not be validated.
                  details: {}
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            reflection_body_empty:
              summary: Empty or whitespace-only Reflection
              value:
                error:
                  code: validation_error
                  message: The request could not be validated.
                  details:
                    body: empty
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
            reflection_body_too_long:
              summary: Reflection exceeds character limit
              value:
                error:
                  code: validation_error
                  message: The request could not be validated.
                  details:
                    body: too_long
                  request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
    InternalError:
      description: Unexpected server error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal_error
              message: An unexpected error occurred.
              details: {}
              request_id: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
...
  schemas:
    UUID:
      type: string
      format: uuid
      description: Canonical UUID string identifier.
      example: f7d7c950-2892-4b6c-9300-ef6c5cbcb2d1
...
    PublicUser:
      type: object
      description: Public-safe user projection. Excludes private and moderation fields.
      required:
        - id
        - username
        - display_name
        - bio
        - avatar_url
        - submission_count
        - current_streak
        - is_self
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        username:
          type: string
          example: sketchy_matt
        display_name:
          type: string
          example: Matt
        bio:
          type:
            - string
            - "null"
          example: Sketching every day.
        avatar_url:
          type:
            - string
            - "null"
          description: Signed or CDN avatar URL when available; null until avatar upload exists.
          format: uri
          example: null
        submission_count:
          type: integer
          minimum: 0
          description: Count of published, non-deleted Submissions on this profile.
          example: 12
        current_streak:
          type: integer
          minimum: 0
          description: |
            Consecutive Prompt Dates (UTC) with at least one published
            Submission, ending today or yesterday. Multiple Submissions on
            the same date count once.
          example: 5
        is_self:
          type: boolean
          description: True when the authenticated viewer is this profile owner.
          example: false
...
    RecentFeed:
      type: object
      description: Cursor-paginated recent community feed page.
      required:
        - items
        - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FeedItem"
        next_cursor:
          type:
            - string
            - "null"
          description: Opaque cursor for the next page, or null when no further pages exist.
          example: null
...
    CreativeType:
      type: string
      description: The type of creative activity.
      enum:
        - sketch
        - story
...
    SubmissionVisibility:
      type: string
      description: Submission visibility. Version one supports public only.
      enum:
        - public
    SubmissionStatus:
      type: string
      description: Submission lifecycle status.
      enum:
        - processing
        - published
        - hidden
        - removed
        - deleted
...
    CreateSubmissionRequest:
      type: object
      description: |
        Publish a creative submission. Provide the creative type, owned session
        id, and type-specific content. The content discriminator determines
        validation rules.
      additionalProperties: false
      required:
        - creative_type
        - session_id
        - content
      properties:
        creative_type:
          $ref: "#/components/schemas/CreativeType"
        session_id:
          $ref: "#/components/schemas/UUID"
        content:
          oneOf:
            - $ref: "#/components/schemas/SketchSubmissionContent"
            - $ref: "#/components/schemas/StorySubmissionContent"
          discriminator:
            propertyName: creative_type
            mapping:
              sketch: "#/components/schemas/SketchSubmissionContent"
              story: "#/components/schemas/StorySubmissionContent"
    Submission:
      type: object
      description: Public Submission projection with signed media URLs.
      required:
        - id
        - creative_type
        - caption
        - visibility
        - status
        - timer_mode
        - timer_seconds
        - like_count
        - reflection_count
        - viewer_has_liked
        - is_owner
        - image_url
        - thumbnail_url
        - user
        - prompt
        - session_id
        - upload_id
        - body
        - published_at
        - created_at
        - updated_at
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        creative_type:
          $ref: "#/components/schemas/CreativeType"
        caption:
          type:
            - string
            - "null"
          example: A quiet coffee and banana sketch.
        visibility:
          $ref: "#/components/schemas/SubmissionVisibility"
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        timer_mode:
          $ref: "#/components/schemas/TimerMode"
        timer_seconds:
          type:
            - integer
            - "null"
          description: Selected countdown seconds, or null for no_timer.
          example: 300
        like_count:
          type: integer
          minimum: 0
          example: 0
        reflection_count:
          type: integer
          minimum: 0
          example: 0
        viewer_has_liked:
          type: boolean
          example: false
        is_owner:
          type: boolean
          example: true
        image_url:
          type:
            - string
            - "null"
          format: uri
          description: Short-lived signed display image URL.
        thumbnail_url:
          type:
            - string
            - "null"
          format: uri
          description: Short-lived signed thumbnail URL.
        user:
          $ref: "#/components/schemas/FeedUserSummary"
        prompt:
          $ref: "#/components/schemas/FeedPromptSummary"
        session_id:
          $ref: "#/components/schemas/UUID"
        upload_id:
          oneOf:
            - $ref: "#/components/schemas/UUID"
            - type: "null"
        body:
          type:
            - string
            - "null"
          description: Full story body text. Null for non-story submissions.
        published_at:
          $ref: "#/components/schemas/Timestamp"
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
...