# backend/app/api/v1/users.py
"""Public user profile routes."""

from __future__ import annotations

from fastapi import APIRouter, Depends, Query
from sqlalchemy.ext.asyncio import AsyncSession

from app.auth.deps import get_optional_current_user
from app.core.clock import Clock, get_clock
from app.core.settings import Settings, get_settings
from app.db.session import get_db_session
from app.models.enums import CreativeType
from app.models.user import User
from app.schemas.feed import RecentFeedResponse
from app.schemas.me import PublicUserResponse
from app.services.profile import ProfileService
from app.storage.base import StorageAdapter, get_storage_adapter

router = APIRouter(tags=["users"])


@router.get("/users/{username}", response_model=PublicUserResponse)
async def get_public_user(
    username: str,
    creative_type: CreativeType = Query(...),
    viewer: User | None = Depends(get_optional_current_user),
    session: AsyncSession = Depends(get_db_session),
    clock: Clock = Depends(get_clock),
    storage: StorageAdapter = Depends(get_storage_adapter),
    settings: Settings = Depends(get_settings),
) -> PublicUserResponse:
    return await ProfileService(
        session,
        clock=clock,
        storage=storage,
        settings=settings,
    ).get_public_user(username, viewer=viewer, creative_type=creative_type)


@router.get("/users/{username}/submissions", response_model=RecentFeedResponse)
async def get_user_submissions(
    username: str,
    cursor: str | None = Query(default=None),
    limit: int = Query(default=20, ge=1, le=50),
    creative_type: CreativeType = Query(...),
    viewer: User | None = Depends(get_optional_current_user),
    session: AsyncSession = Depends(get_db_session),
    clock: Clock = Depends(get_clock),
    storage: StorageAdapter = Depends(get_storage_adapter),
    settings: Settings = Depends(get_settings),
) -> RecentFeedResponse:
    return await ProfileService(
        session,
        clock=clock,
        storage=storage,
        settings=settings,
    ).get_user_submissions(
        username,
        cursor=cursor,
        limit=limit,
        viewer=viewer,
        creative_type=creative_type,
    )